requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
python-dotenv
playwright
sentence-transformers
httpx
